2_demo_via_http_2/requirements.txt
Flask>=2.0,<3.0
Werkzeug>=2.0,<3.0
google-cloud-pubsub==2.15.1
